Bug fix learning

+++ Domain/Entity/Models/Learning/CourseItem.cs
@@ -5,7 +5,7 @@ using Entity.Models.Common;
 namespace Entity.Models.Learning;
 
 [Table("course_items", Schema = "learning")]
-public class CourseItem : AuditableModelBase<int>
+public class CourseItem : AuditableModelBase<long>
 {
     [Column("title")]
     public string Title { get; set; }
@@ -27,24 +27,24 @@ public class CourseItem : AuditableModelBase<int>
 
     [Column("module_id")]
     [ForeignKey(nameof(Module))]
-    public int ModuleId { get; set; }
+    public long ModuleId { get; set; }
 
     public virtual Module Module { get; set; }
 
     [Column("quiz_id")]
     [ForeignKey(nameof(Quiz))]
-    public int? QuizId { get; set; }
+    public long? QuizId { get; set; }
 
     public virtual Quiz? Quiz { get; set; }
 
     [Column("video_of_course_id")]
     [ForeignKey(nameof(VideoOfCourse))]
-    public int? VideoOfCourseId { get; set; }
+    public long? VideoOfCourseId { get; set; }
     public virtual VideoOfCourse? VideoOfCourse { get; set; }
 
     [Column("text_content_id")]
     [ForeignKey(nameof(TextContent))]
-    public int? TextContentId { get; set; }
+    public long? TextContentId { get; set; }
 
     public virtual TextContent? TextContent { get; set; }
 }
